Move to a new bare repository
Why are we doing this?

Well, the problem with stow was that when other programs would bind
mount files like flatpak, it would symlink from a symlink, which would
work the first time and then fail after that. That's lame.

To avoid symlink to symlink issues, we instead use the real files in our
real folder locations. Thus bind mounts and other weird symlinking
should work, as expected, each and every time.