Update container for adguardhome work with podman
On MacOS at least, podman does not like binding /work session.db
for some reason. Luckily, this is only for logs, so for now we
can just make a "new" db every time instead of bind mounting one.

CAPS are needed for some reason all of a sudden in order for stuff to
run

Must run in a rootful podman-machine.
The podman-machine itself must have the systemd-resolved stub listener
turned off