Move off flatpak to jail
Normal applications are safer in some cases

Many flatpaks are distributed with extra self-packaged libs
outside of the runtime

We should only accept flatpaks which use only bits from the runtime.

So far, the only app that seems to pass that test is g4music.

Moonlight is still used as a flatpak because Shortcut Inhibit does not
work on Gnome Wayland outside of flatpak.