Don't mess with iptables-legacy rules
When you set this, the container runs `iptables-legacy` which messes
up the firewall on our system, as the system runs `nftables` and
`firewalld`.

Running two firewalls causes weird issues and we want to avoid this
problem.

So instead, you could set TS_DEBUG_FIREWALL_MODE=nftables and force
Tailscale to run nftables rules, and this works, EXCEPT that it then
breaks `iptables-nft` because of some issue with the `mask` rule applied
in the `ts-forward` chain.

Breaking the rule affects other containers that run as root and require
netardvk, like `crowdsec` because that bad mask rule invalidates the
whole chain of rules, and causes those containers to crash.

----

Instead, we turn OFF the netfilter rule mode, and we manually specify
our tailscale rules as a service in firewalld, and open the related UDP
port

This allows us to not run double firewalls, not break iptables-nft, and
still allow Tailscale direct connections instead of falling back to
relay DERP